requirements.txt
discord.py==2.3.2
anthropic==0.28.0
gTTS==2.5.1
httpx==0.27.0
